+++ src/main/java/com/textbasedgame/world/shopitems.java
@@ -13,12 +13,10 @@ import com.textbasedgame.items.legsArmorItems.*;
 import com.textbasedgame.items.handItems.*;
 import com.textbasedgame.playerFiles.player;
 import com.textbasedgame.util.TrekkerMath;
-import com.textbasedgame.util.itemInfoPrinter;
-import com.textbasedgame.util.response;
 
 public abstract class shopitems {
     ////////////////////////////////////////////////////////////////////////////////////////////////
-    // precondition: all shops are 4 items
+    // precondition: all shops are 6 items
     ////////////////////////////////////////////////////////////////////////////////////////////////
     public static ArrayList<Class<? extends item>> allItemsList = new ArrayList<>();
 
@@ -27,7 +25,7 @@ public abstract class shopitems {
     
     public static int[] itemPrice = {5, 1, 25, 30, 10, 3, 30};
     
-    private static item[] itemsInShop = new item[4];
+    private static item[] itemsInShop = new item[6];
 
     public static void createShopItemsArr(){
         consumableShopItems.add(fish.class);
@@ -163,9 +161,6 @@ public abstract class shopitems {
 
         item toAdd = itemsInShop[shopItemNum - 1];
     
-        buyConfirmationMenu(toAdd);
-
-
         player.addItemToPlayer(toAdd);
         
 
@@ -174,7 +169,7 @@ public abstract class shopitems {
         createShop();
     }
     public static void buyItem(item i){
-    
+
         player.addItemToPlayer(i);
         
 
@@ -199,30 +194,7 @@ public abstract class shopitems {
         return null;
     }
 
-    private static boolean buyConfirmationMenu(item i){
-
-        if(i instanceof consumables){
-            itemInfoPrinter.printConsumableInfo((consumables)i);
-        }
-        else if(i instanceof equipables){
-            itemInfoPrinter.printEquipablesInfo((equipables) i);
-        }
-
-        gui.printOnGameSide("-------------------------------------");
-        gui.newlOnGameSide();
-        gui.printOnGameSide("Are you sure you want to buy " + i.getItemName() + " for " + i.getPrice() + " shmeckles?");
-        gui.printOnGameSide("1: Yes");
-        gui.printOnGameSide("2: No");
-
-        String input = gui.getInput();
-
-        if(input.equals("1") || response.respondYes(input)){
-            return true;
-        }
-        else{
-            return false;
-        }
-    }
+    
 
 
 }